# PERF: optimize anynan/allnan with transpose axis